Rename AddCommand to Commands::Add
